+++ internal/src/util/auth.go
@@ -13,7 +13,7 @@ const UserEmail = "user_email"
 type AuthKey struct{}
 
 type UserInfo struct {
-	UserID    uint
+	UserID    int64
 	UserName  string
 	UserEmail string
 }
